Add more verifications, reworked some files

+++ controller/adminController.php
@@ -18,43 +18,76 @@ function admin_showPanel()
 {
     $userManager = new UserManager();
     $userList = $userManager->getAllUser();
-
     $commentManager = new CommentManager();
     $commentList = $commentManager->getAllReportComment();
-
     $postManager = new PostManager();
     $postList = $postManager->getAllPostAdmin();
-
     $title = "Administration";
-
     require('view/backend/admin.php');
 }
 
 function unReportComment($commentId)
 {
-    $title = "Administration";
+    if(isset($_SESSION['user']))
+    {
+        if($_SESSION['user']->getGroupId() != User::IS_USER)
+        {
+            $title = "Administration";
 
-    $commentManager = new CommentManager();
-    $comment = $commentManager->getCommentById($commentId);
+            $commentManager = new CommentManager();
+            $comment = $commentManager->getCommentById($commentId);
 
-    $comment->setReports(0);
+            $comment->setReports(0);
 
-    $commentManager->updateComment($comment);
+            $commentManager->updateComment($comment);
 
-    header('Location:index.php?action=showAccount');
+            header('Location:index.php?action=showAccount');
+        }
+        else
+        {
+            showError404('Vous n\'êtes pas autorisé à effectuer cette action');
+        }
+    }
+    else
+    {
+        showError404('Vous devez être connecté pour effectuer cette action');
+    }
 }
 
 function editUserComment($commentId)
 {
-    $title = 'Modifier le commentaire';
-    $commentManager = new CommentManager();
-    $comment = $commentManager->getCommentById($commentId);
+    if(isset($_SESSION['user']))
+    {
+        if($_SESSION['user']->getGroupId() != User::IS_USER)
+        {
+            $commentManager = new CommentManager();
+            $comment = $commentManager->getCommentById($commentId);
+            $title = 'Modifier le commentaire';
 
-    require('view/frontend/comment-edit.php');
+            if(is_null($comment))
+            {
+                showError404('Le commentaire recherché est introuvable');
+            }
+            else
+            {
+                require('view/frontend/comment-edit.php');
+            }
+        }
+        else
+        {
+            showError404('Vous n\'êtes pas autorisé à effectuer cette action');
+        }
+    }
+    else
+    {
+        showError404('Vous devez être connecté pour effectuer cette action');
+    }
+
 }
 
 function sendUserEditComment($datas)
 {
+    $title = 'Modifier le commentaire';
     $commentManager = new CommentManager();
     $comment = $commentManager->getCommentById($datas['commentId']);
 
@@ -91,46 +124,76 @@ function sendUserEditComment($datas)
     }
     else
     {
-        $message = 'Veuillez rentrer un commentaire';
+        $message = '<div class="alert alert-danger">Veuillez rentrer un commentaire</div>';
         return require('view/frontend/comment-edit.php');
     }
 }
 
 function deleteUserComment($commentId)
 {
-    $commentManager = new CommentManager();
-    $comment = $commentManager->getCommentById($commentId);
+    if(isset($_SESSION['user']))
+    {
+        if($_SESSION['user']->getGroupId() != User::IS_USER)
+        {
+            $commentManager = new CommentManager();
+            $comment = $commentManager->getCommentById($commentId);
 
-    $commentManager->deleteComment($comment);
+            if(is_null($comment))
+            {
+                showError404('Le commentaire recherché est introuvable');
+            }
+            else
+            {
+                $commentManager->deleteComment($comment);
+                $userManager = new UserManager();
+                $userList = $userManager->getAllUser();
+                $commentManager = new CommentManager();
+                $commentList = $commentManager->getAllReportComment();
+                $postManager = new PostManager();
+                $postList = $postManager->getAllPost(5,0);
+                $title = "Administration";
+                header('Location:index.php?action=showAccount');
+            }
+        }
+        else
+        {
+            showError404('Vous n\'êtes pas autorisé à accéder à cette page');
+        }
+    }
+    else
+    {
+        showError404('Vous devez être connecté pour accéder à cette page');
+    }
 
-    $userManager = new UserManager();
-    $userList = $userManager->getAllUser();
-
-    $commentManager = new CommentManager();
-    $commentList = $commentManager->getAllReportComment();
-
-    $postManager = new PostManager();
-    $postList = $postManager->getAllPost(5,0);
-
-    $title = "Administration";
-
-    header('Location:index.php?action=showAccount');
 }
 
 function editUser($userId)
 {
-
-    $userManager = new UserManager();
-    $user = $userManager->getUserByNameOrId($userId);
-
-    if(is_null($user))
+    if(isset($_SESSION['user']))
     {
-        return showError404('L\'utilisateur recherché n\'existe pas');
+        if($_SESSION['user']->getGroupId() != User::IS_USER)
+        {
+            $userManager = new UserManager();
+            $user = $userManager->getUserByNameOrId($userId);
+
+            if(is_null($user))
+            {
+                showError404('L\'utilisateur recherché n\'existe pas');
+            }
+            else
+            {
+                $title = 'Edition de l\'utilisateur : ' . $user->getUsername() .'';
+                require('view/backend/user-edit.php');
+            }
+        }
+        else
+        {
+            showError404('Vous n\'êtes pas autorisé à accéder à cette page');
+        }
     }
     else
     {
-        $title = 'Edition de l\'utilisateur : ' . $user->getUsername() .'';
-        require('view/backend/user-edit.php');
+        showError404('Vous devez être connecté pour accéder à cette page');
     }
 }
 
@@ -167,24 +230,48 @@ function sendUserEdit($datas)
 
 function promoteUser($userId, $groupId)
 {
-    $userManager = new UserManager();
-    $user = $userManager->getUserByNameOrId($userId);
+    if(isset($_SESSION['user']))
+    {
+        if($_SESSION['user']->getGroupId() != User::IS_USER)
+        {
+            $userManager = new UserManager();
+            $user = $userManager->getUserByNameOrId($userId);
+            $user->setGroupId($groupId);
+            $userManager->updateUser($user);
+            header('Location:index.php?action=showAccount');
+        }
+        else
+        {
+            showError404('Vous n\'êtes pas autorisé à effectuer cette action');
+        }
+    }
+    else
+    {
+        showError404('Vous devez être connecté pour effectuer cette action');
+    }
 
-    $user->setGroupId($groupId);
-
-    $userManager->updateUser($user);
-
-    header('Location:index.php?action=showAccount');
 }
 
 function deleteUser($userId)
 {
-    $userManager = new UserManager();
-    $user = $userManager->getUserByNameOrId($userId);
-
-    $userManager->deleteUser($user);
-
-    header('Location:index.php?action=showAccount');
+    if(isset($_SESSION['user']))
+    {
+        if($_SESSION['user']->getGroupId() != User::IS_USER)
+        {
+            $userManager = new UserManager();
+            $user = $userManager->getUserByNameOrId($userId);
+            $userManager->deleteUser($user);
+            header('Location:index.php?action=showAccount');
+        }
+        else
+        {
+            showError404('Vous n\'êtes pas autorisé à effectuer cette action');
+        }
+    }
+    else
+    {
+        showError404('Vous devez être connecté pour effectuer cette action');
+    }
 }
 
 /**
@@ -196,62 +283,72 @@ function createPost($datas)
     $title = 'Creation d\'article';
     $postManager = new PostManager();
 
-
-    if(!empty($datas['content']) && !empty($datas['postRecap']) && !empty($datas['title']))
+    if(isset($_SESSION['user']) && $_SESSION['user']->getGroupId() == User::IS_AUTHOR)
     {
-        $postArray = array(
-            'title' => $datas['title'],
-            'postRecap' => $datas['postRecap'],
-            'content' => $datas['content'],
-
-        );
-
-        $post = new Post($postArray);
-
-        if(preg_match('/(<script>)/', $datas['title']) || preg_match('/(<script>)/', $datas['postRecap']) || preg_match('/(<script>)/', $datas['content']))
+        if(!empty($datas['content']) && !empty($datas['postRecap']) && !empty($datas['title']))
         {
-            $message = '<div class="alert alert-danger">L\'insertion de script n\'est pas autorisée, pour plus d\'information, consultez le développeur</div>';
+            $postArray = array(
+                'title' => $datas['title'],
+                'postRecap' => $datas['postRecap'],
+                'content' => $datas['content'],
 
-            return require('view/backend/post-add.php');
-        }
+            );
 
-        $postManager->createPost($post);
+            $post = new Post($postArray);
 
-        $userManager = new UserManager();
-        $userList = $userManager->getAllUser();
+            if(preg_match('/(<script>)/', $datas['title']) || preg_match('/(<script>)/', $datas['postRecap']) || preg_match('/(<script>)/', $datas['content']))
+            {
+                $message = '<div class="alert alert-danger">L\'insertion de script n\'est pas autorisée, pour plus d\'information, consultez le développeur</div>';
 
-        $commentManager = new CommentManager();
-        $commentList = $commentManager->getAllReportComment();
+                return require('view/backend/post-add.php');
+            }
 
-        $postManager = new PostManager();
-        $postList = $postManager->getAllPost(5,0);
+            $postManager->createPost($post);
 
-        $title = "Administration";
+            $userManager = new UserManager();
+            $userList = $userManager->getAllUser();
 
-        $message = '<div class="alert alert-success alert-dismissible fade show" role="alert">
+            $commentManager = new CommentManager();
+            $commentList = $commentManager->getAllReportComment();
+
+            $postManager = new PostManager();
+            $postList = $postManager->getAllPost(5,0);
+
+            $title = "Administration";
+
+            $message = '<div class="alert alert-success alert-dismissible fade show" role="alert">
           <strong>Votre Article a bien été créé !</strong>
           <button type="button" class="close" data-dismiss="alert" aria-label="Close">
             <span aria-hidden="true">&times;</span>
           </button>
         </div>';
 
-        require('view/backend/admin.php');
+            require('view/backend/admin.php');
+        }
+        else
+        {
+            require('view/backend/post-add.php');
+        }
     }
     else
     {
-        require('view/backend/post-add.php');
+      showError404('Vous n\'êtes pas autorisé à accéder à cette page.');
     }
-
-
 }
 
 function showPostEdit($postId) {
 
-    $title = 'Edition de l\'article';
-    $postManager = new PostManager();
-    $post = $postManager->getPostById($postId);
-
-    require('view/backend/post-edit.php');
+    if(isset($_SESSION['user']) && $_SESSION['user']->getGroupId() == User::IS_AUTHOR)
+    {
+        $title = 'Edition de l\'article';
+        $postManager = new PostManager();
+        $post = $postManager->getPostById($postId);
+        require('view/backend/post-edit.php');
+    }
+    else
+    {
+        showError404('Vous n\'êtes pas autorisé à accéder à cette page');
+    }
 }
 
 function sendPostEdit($datas)
@@ -288,11 +385,18 @@ function sendPostEdit($datas)
 
 function deletePost($postId)
 {
-    $postManager = new PostManager();
-    $post = $postManager->getPostById($postId);
+    if(isset($_SESSION['user']) && $_SESSION['user']->getGroupId() == User::IS_AUTHOR)
+    {
+        $postManager = new PostManager();
+        $post = $postManager->getPostById($postId);
 
-    $postManager->deletePost($post);
+        $postManager->deletePost($post);
 
-    header('Location:index.php?action=showAccount');
+        header('Location:index.php?action=showAccount');
+    }
+    else
+    {
+        showError404('Vous n\'êtes pas autorisé à accéder à cette page');
+    }
 }
